update accu image for long

+++ EventMessageVisualisationStreamer/src/main/java/Image/AccumulatedImagePOJO.java
@@ -47,14 +47,14 @@ public class AccumulatedImagePOJO implements ImageInterface {
         return image;
     }
 
-    public int getFrequency(int detector) {
+    public long getFrequency(long detector) {
         if (!image.containsKey(detector)) {
             throw new InvalidParameterException(MISSING_KEY_ERROR_MESSAGE);
         }
         return (int) image.get(detector);
     }
 
-    public void setFrequency(int detector, int newFreq) {
+    public void setFrequency(long detector, long newFreq) {
         if (detector < 0) {
             throw new InvalidParameterException(DETECTOR_ID_POSITIVE_ERROR_MESSAGE);
         } else if (newFreq < 0) {
@@ -63,7 +63,7 @@ public class AccumulatedImagePOJO implements ImageInterface {
         image.put(detector,newFreq);
     }
 
-    public void incrementFrequency(int detector) {
+    public void incrementFrequency(long detector) {
         int oldFreq = 0;
         if (image.containsKey(detector)) {
             oldFreq = (int) image.get(detector);
@@ -85,7 +85,7 @@ public class AccumulatedImagePOJO implements ImageInterface {
 
         for (Object detector: frameImage.getImage().keySet()) {
             int detectorId = (int) detector;
-            int newFreq = frameImage.getFrequency(detectorId);
+            long newFreq = frameImage.getFrequency(detectorId);
             if (this.getImage().containsKey(detectorId)) {
                 newFreq += this.getFrequency(detectorId);
             }
